requirements.txt
django
djangorestframework
django-cors-headers
yarn
dotenv